# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Runtime 示例应用 API 文档。",
        "title": "Agent Runtime API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "post": {
                "description": "创建一个新的后台任务，并立即返回任务快照。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "创建任务",
                "parameters": [
                    {
                        "description": "创建任务请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/tasks/running": {
            "get": {
                "description": "根据 conversation id 返回最近一个非终态任务；若不存在则返回 null。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "查询会话运行中任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "conversation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "根据任务 id 返回当前任务快照。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "获取任务详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/cancel": {
            "post": {
                "description": "发起任务取消请求；若任务尚未执行，会直接进入 cancelled。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "取消任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/events": {
            "get": {
                "description": "先补发 after_seq 之后的历史事件，再持续推送实时事件。",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "订阅任务事件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "仅返回该序号之后的事件",
                        "name": "after_seq",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE event stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/retry": {
            "post": {
                "description": "基于一个已存在任务创建新的重试任务，并返回新的任务快照。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "重试任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TaskSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/workspace/confirm": {
            "post": {
                "description": "将 pending_merge 状态的 mutable task workspace 备份并整目录回写到用户 home workspace。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "确认合并任务工作区",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WorkspaceStateSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/workspace/discard": {
            "post": {
                "description": "将 task workspace 标记为 discarded，保留目录但不回写用户 home workspace。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "放弃任务工作区变更",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WorkspaceStateSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorSwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "created_by": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "task_type": {
                    "type": "string"
                },
                "workspace_mode": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TaskSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/handlers.TaskSwaggerDoc"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "handlers.WorkspaceStateSwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/handlers.WorkspaceStateSwaggerDoc"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        }
    }
}